requirements.txt
alembic==1.13.2
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
dash==2.18.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
extensions==0.4
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
frozendict==2.4.4
greenlet==3.1.0
html5lib==1.1
idna==3.8
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.3.0
Mako==1.3.5
MarkupSafe==2.1.5
multitasking==0.0.11
nest-asyncio==1.6.0
numpy==2.1.1
packaging==24.1
pandas==2.2.2
peewee==3.17.6
platformdirs==4.3.2
plotly==5.24.1
psycopg2-binary==2.9.9
pygame==2.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
retrying==1.3.4
six==1.16.0
soupsieve==2.6
SQLAlchemy==2.0.34
tenacity==9.0.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
webencodings==0.5.1
Werkzeug==3.0.4
yfinance==0.2.43
zipp==3.20.2
gunicorn